update cpp tests
