+++ tests/common.cpp
@@ -1,0 +1,165 @@
+/*******************************************************************************
+*   (c) 2018 - 2023 Zondax AG
+*
+*  Licensed under the Apache License, Version 2.0 (the "License");
+*  you may not use this file except in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+*  Unless required by applicable law or agreed to in writing, software
+*  distributed under the License is distributed on an "AS IS" BASIS,
+*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+*  See the License for the specific language governing permissions and
+*  limitations under the License.
+********************************************************************************/
+#include "gmock/gmock.h"
+#include "parser_impl.h"
+
+#include <parser.h>
+#include <sstream>
+#include <string>
+#include <fmt/core.h>
+#include "common.h"
+#include <iostream>
+#include <fstream>
+#include <json/json.h>
+#include <app_mode.h>
+#include <hexutils.h>
+#include "parser.h"
+
+std::vector<std::string> dumpUI(parser_context_t *ctx,
+                                uint16_t maxKeyLen,
+                                uint16_t maxValueLen) {
+    auto answer = std::vector<std::string>();
+
+    uint8_t numItems;
+    parser_error_t err = parser_getNumItems(ctx, &numItems);
+    if (err != parser_ok) {
+        return answer;
+    }
+
+    for (uint8_t idx = 0; idx < numItems; idx++) {
+        char keyBuffer[1000];
+        char valueBuffer[1000];
+        uint8_t pageIdx = 0;
+        uint8_t pageCount = 1;
+
+        while (pageIdx < pageCount) {
+            std::stringstream ss;
+
+            err = parser_getItem(ctx, idx,
+                                 keyBuffer, maxKeyLen,
+                                 valueBuffer, maxValueLen,
+                                 pageIdx, &pageCount);
+
+            ss << fmt::format("{} | {}", idx, keyBuffer);
+            if (pageCount > 1) {
+                ss << fmt::format(" [{}/{}]", pageIdx + 1, pageCount);
+            }
+            ss << " : ";
+
+            if (err == parser_ok) {
+                // Model multiple lines
+                ss << fmt::format("{}", valueBuffer);
+            } else {
+                ss << parser_getErrorDescription(err);
+            }
+
+            auto output = ss.str();
+            answer.push_back(output);
+
+            pageIdx++;
+        }
+    }
+
+    return answer;
+}
+
+std::string CleanTestname(std::string s) {
+    s.erase(remove_if(s.begin(), s.end(), [](char v) -> bool {
+        return v == ':' || v == ' ' || v == '/' || v == '-' || v == '.' || v == '_' || v == '#';
+    }), s.end());
+    return s;
+}
+
+// Retrieve testcases from json file
+std::vector<testcase_t> GetJsonTestCases(const std::string &jsonFile) {
+    auto answer = std::vector<testcase_t>();
+
+    Json::CharReaderBuilder builder;
+    Json::Value obj;
+
+    std::string fullPathJsonFile = std::string(TESTVECTORS_DIR) + jsonFile;
+
+    std::ifstream inFile(fullPathJsonFile);
+    if (!inFile.is_open()) {
+        return answer;
+    }
+
+    // Retrieve all test cases
+    JSONCPP_STRING errs;
+    Json::parseFromStream(builder, inFile, &obj, &errs);
+    std::cout << "Number of testcases: " << obj.size() << std::endl;
+
+    for (int i = 0; i < obj.size(); i++) {
+
+        auto outputs = std::vector<std::string>();
+        for (auto s : obj[i]["output"]) {
+            outputs.push_back(s.asString());
+        }
+
+        auto outputs_expert = std::vector<std::string>();
+        for (auto s : obj[i]["output_expert"]) {
+            outputs_expert.push_back(s.asString());
+        }
+
+        answer.push_back(testcase_t{
+                obj[i]["index"].asUInt64(),
+                obj[i]["name"].asString(),
+                obj[i]["blob"].asString(),
+                outputs,
+                outputs_expert
+        });
+    }
+
+    return answer;
+}
+
+void check_testcase(const testcase_t &tc, bool expert_mode) {
+    app_mode_set_expert(expert_mode);
+
+    parser_context_t ctx = {0};
+    parser_error_t err = parser_unexpected_error;
+
+    uint8_t buffer[10000] = {0};
+    const uint16_t bufferLen = parseHexString(buffer, sizeof(buffer), tc.blob.c_str());
+
+    parser_tx_t tx_obj;
+    memset(&tx_obj, 0, sizeof(tx_obj));
+
+    err = parser_parse(&ctx, buffer, bufferLen, &tx_obj);
+    ASSERT_EQ(err, parser_ok) << parser_getErrorDescription(err);
+
+    err = parser_validate(&ctx);
+    ASSERT_EQ(err, parser_ok) << parser_getErrorDescription(err);
+
+
+    auto output = dumpUI(&ctx, 39, 39);
+
+    std::cout << std::endl;
+    for (const auto &i : output) {
+        std::cout << i << std::endl;
+    }
+    std::cout << std::endl << std::endl;
+
+#if 1
+    std::vector<std::string> expected = app_mode_expert() ? tc.expected_expert : tc.expected;
+    EXPECT_EQ(output.size(), expected.size());
+    for (size_t i = 0; i < expected.size(); i++) {
+        if (i < output.size()) {
+            EXPECT_THAT(output[i], testing::Eq(expected[i]));
+        }
+    }
+#endif
+}
