+++ tests/ui_tests.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2018 - 2022 Zondax AG
+*   (c) 2018 - 2023 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -15,116 +15,12 @@
 ********************************************************************************/
 
 #include "gmock/gmock.h"
-
-#include <iostream>
-#include <fstream>
-#include <json/json.h>
-#include <hexutils.h>
-#include <parser_txdef.h>
-#include "parser.h"
-#include "app_mode.h"
-#include "utils/common.h"
+#include "common.h"
 
 using ::testing::TestWithParam;
 
-typedef struct {
-    uint64_t index;
-    std::string name;
-    std::string blob;
-    std::vector<std::string> expected;
-    std::vector<std::string> expected_expert;
-} testcase_t;
-
-class JsonTestsA : public ::testing::TestWithParam<testcase_t> {
-public:
-    struct PrintToStringParamName {
-        template<class ParamType>
-        std::string operator()(const testing::TestParamInfo<ParamType> &info) const {
-            auto p = static_cast<testcase_t>(info.param);
-            std::stringstream ss;
-            ss << p.index << "_" << p.name;
-            return ss.str();
-        }
-    };
-};
-
-// Retrieve testcases from json file
-std::vector<testcase_t> GetJsonTestCases(std::string jsonFile) {
-    auto answer = std::vector<testcase_t>();
-
-    Json::CharReaderBuilder builder;
-    Json::Value obj;
-
-    std::string fullPathJsonFile = std::string(TESTVECTORS_DIR) + jsonFile;
-
-    std::ifstream inFile(fullPathJsonFile);
-    if (!inFile.is_open()) {
-        return answer;
-    }
-
-    // Retrieve all test cases
-    JSONCPP_STRING errs;
-    Json::parseFromStream(builder, inFile, &obj, &errs);
-    std::cout << "Number of testcases: " << obj.size() << std::endl;
-
-    for (int i = 0; i < obj.size(); i++) {
-
-        auto outputs = std::vector<std::string>();
-        for (auto s : obj[i]["output"]) {
-            outputs.push_back(s.asString());
-        }
-
-        auto outputs_expert = std::vector<std::string>();
-        for (auto s : obj[i]["output_expert"]) {
-            outputs_expert.push_back(s.asString());
-        }
-
-        answer.push_back(testcase_t{
-                obj[i]["index"].asUInt64(),
-                obj[i]["name"].asString(),
-                obj[i]["blob"].asString(),
-                outputs,
-                outputs_expert
-        });
-    }
-
-    return answer;
-}
-
-void check_testcase(const testcase_t &tc, bool expert_mode) {
-
-    app_mode_set_expert(expert_mode);
-
-    parser_context_t ctx;
-    parser_error_t err;
-
-    uint8_t buffer[10000];
-    uint16_t bufferLen = parseHexString(buffer, sizeof(buffer), tc.blob.c_str());
-
-    parser_tx_t tx_obj;
-    memset(&tx_obj, 0, sizeof(tx_obj));
-
-    err = parser_parse(&ctx, buffer, bufferLen, &tx_obj);
-
-    ASSERT_EQ(err, parser_ok) << parser_getErrorDescription(err);
-
-    auto output = dumpUI(&ctx, 39, 39);
-
-    std::cout << std::endl;
-    for (const auto &i : output) {
-        std::cout << i << std::endl;
-    }
-    std::cout << std::endl << std::endl;
-
-    std::vector<std::string> expected = app_mode_expert() ? tc.expected_expert : tc.expected;
-    EXPECT_EQ(output.size(), expected.size());
-    for (size_t i = 0; i < expected.size(); i++) {
-        if (i < output.size()) {
-            EXPECT_THAT(output[i], testing::Eq(expected[i]));
-        }
-    }
-}
-
+class JsonTestsA : public JsonTests_Base {};
+class JsonTestsB : public JsonTests_Base {};
 
 INSTANTIATE_TEST_SUITE_P
 (
@@ -135,3 +31,14 @@ INSTANTIATE_TEST_SUITE_P
 );
 TEST_P(JsonTestsA, CheckUIOutput_CurrentTX_Normal) { check_testcase(GetParam(), false); }
 TEST_P(JsonTestsA, CheckUIOutput_CurrentTX_Expert) { check_testcase(GetParam(), true); }
+
+
+INSTANTIATE_TEST_SUITE_P
+(
+    JsonTestCasesCurrentTxVer,
+    JsonTestsB,
+    ::testing::ValuesIn(GetJsonTestCases("ibc.json")),
+    JsonTestsB::PrintToStringParamName()
+);
+TEST_P(JsonTestsB, CheckUIOutput_CurrentTX_Normal) { check_testcase(GetParam(), false); }
+TEST_P(JsonTestsB, CheckUIOutput_CurrentTX_Expert) { check_testcase(GetParam(), true); }
