--- tests/utils/common.h
@@ -1,21 +1,0 @@
-/*******************************************************************************
-*   (c) 2018 - 2022 Zondax AG
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-********************************************************************************/
-#include <vector>
-#include <string>
-
-std::vector<std::string> dumpUI(parser_context_t *ctx,
-                                uint16_t maxKeyLen,
-                                uint16_t maxValueLen);
